Er, I meant all map/unmap problems in that last log.
Added -nostdin option, so that the console doesn't select on fd 0
when you don't want it to. When console is started from dash (which
itself forks into the background) and tries to select on its
stdin, Solaris returns an ENXIO (6: No such device or address) as
undocumented in the man page.